move oshdbtag, deprecated rawTag access
