+++ oshdb-api/src/test/java/org/heigit/ohsome/oshdb/api/tests/TestFlatMapReduce.java
@@ -1,7 +1,6 @@
 package org.heigit.ohsome.oshdb.api.tests;
 
 import static org.junit.Assert.assertEquals;
-
 import java.util.AbstractMap.SimpleImmutableEntry;
 import java.util.ArrayList;
 import java.util.HashSet;
@@ -9,6 +8,7 @@ import java.util.List;
 import java.util.Map.Entry;
 import java.util.Set;
 import java.util.TreeSet;
+import org.heigit.ohsome.oshdb.OSHDBTag;
 import org.heigit.ohsome.oshdb.api.db.OSHDBDatabase;
 import org.heigit.ohsome.oshdb.api.db.OSHDBH2;
 import org.heigit.ohsome.oshdb.api.mapreducer.MapReducer;
@@ -50,9 +50,8 @@ public class TestFlatMapReduce {
             return new ArrayList<>();
           }
           List<Entry<Integer, Integer>> ret = new ArrayList<>();
-          int[] tags = contribution.getEntityAfter().getRawTags();
-          for (int i = 0; i < tags.length; i += 2) {
-            ret.add(new SimpleImmutableEntry<>(tags[i], tags[i + 1]));
+          for (OSHDBTag tag : contribution.getEntityAfter().getTags()) {
+            ret.add(new SimpleImmutableEntry<>(tag.getKey(), tag.getValue()));
           }
           return ret;
         })
