+++ oshdb-filter/src/main/java/org/heigit/ohsome/oshdb/filter/TagFilterEquals.java
@@ -1,8 +1,8 @@
 package org.heigit.ohsome.oshdb.filter;
 
+import org.heigit.ohsome.oshdb.OSHDBTag;
 import org.heigit.ohsome.oshdb.osh.OSHEntity;
 import org.heigit.ohsome.oshdb.osm.OSMEntity;
-import org.heigit.ohsome.oshdb.util.OSHDBTag;
 
 /**
  * A tag filter which executes a "key=value" check.
