+++ oshdb/src/main/java/org/heigit/ohsome/oshdb/OSHDBTag.java
@@ -1,12 +1,13 @@
-package org.heigit.ohsome.oshdb.util;
+package org.heigit.ohsome.oshdb;
 
 import java.io.Serializable;
 import java.util.Objects;
 
 public class OSHDBTag implements Serializable {
+    
   private static final long serialVersionUID = 1L;
-  private int key;
-  private int value;
+  private final int key;
+  private final int value;
 
   public OSHDBTag(int key, int value) {
     this.key = key;
@@ -27,15 +28,15 @@ public class OSHDBTag implements Serializable {
 
   @Override
   public boolean equals(Object o) {
-    return o instanceof OSHDBTag &&
-        ((OSHDBTag)o).key == this.key && ((OSHDBTag)o).value == this.value;
+    return o instanceof OSHDBTag 
+        && ((OSHDBTag) o).key == this.key && ((OSHDBTag) o).value == this.value;
   }
 
   @Override
   public int hashCode() {
     return Objects.hash(this.key, this.value);
   }
-
+  
   @Override
   public String toString() {
     return Integer.toString(this.key) + "=" + Integer.toString(this.value);
