+++ oshdb/src/main/java/org/heigit/ohsome/oshdb/osh/OSHEntity.java
@@ -16,6 +16,7 @@ public interface OSHEntity {
 
   OSHDBBoundingBox getBoundingBox();
 
+  @Deprecated
   int[] getRawTagKeys();
 
   boolean hasTagKey(OSHDBTagKey tag);
