+++ oshdb/src/main/java/org/heigit/ohsome/oshdb/osm/OSMEntity.java
@@ -4,8 +4,8 @@ import java.util.Arrays;
 import java.util.Iterator;
 import java.util.stream.IntStream;
 import javax.annotation.Nonnull;
+import org.heigit.ohsome.oshdb.OSHDBTag;
 import org.heigit.ohsome.oshdb.OSHDBTimeable;
-import org.heigit.ohsome.oshdb.util.OSHDBTag;
 import org.heigit.ohsome.oshdb.util.OSHDBTagKey;
 
 public abstract class OSMEntity implements OSHDBTimeable {
@@ -87,6 +87,7 @@ public abstract class OSMEntity implements OSHDBTimeable {
     };
   }
 
+  @Deprecated
   public int[] getRawTags() {
     return tags;
   }
